PWM Control + Controller Support

+++ STM32Project/Core/Inc/controller_conversion.h
@@ -4,8 +4,8 @@
 #include "projectConstants.h"
 #include <stdint.h>
 
-float accel_map(uint8_t controllerInput);
-float steer_map(uint8_t controllerInput);
-float brake_map(uint8_t controllerInput);
+int accel_map(int controllerInput);
+int steer_map(int controllerInput);
+int brake_map(int controllerInput);
 void determine_steering_direction(float calculatedValue);
 #endif
