+++ STM32Project/Core/Src/motor_control.c
@@ -3,7 +3,7 @@
 /// Sends a given value to all motors
 /// - Parameters:
 ///   - dutyCycle: the value being sent to simulate a duty cycle
-void send_input_to_all_motors(float dutyCycle){
+void send_input_to_all_motors(int dutyCycle){
     TIM1->CCR1 = dutyCycle;
     TIM1->CCR2 = dutyCycle;
     TIM1->CCR3 = dutyCycle;
@@ -15,7 +15,7 @@ void send_input_to_all_motors(float dutyCycle){
 /// - Parameters:
 ///   - dutyCycle: the value being sent to simulate a duty cycle
 ///   - motor: the motor who's duty cycle is being changed
-void send_to_motor(float dutyCycle, int motor){
+void send_to_motor(int dutyCycle, int motor){
 	switch(motor){
 		case 0:
 			TIM1->CCR1 = dutyCycle;
